Move logic to new file
