+++ lib/rundoc/code_command/empty_binding.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require "erb"
+
+class EmptyBinding
+  def self.create
+    new.empty_binding
+  end
+
+  def empty_binding
+    binding
+  end
+end
+
+module Rundoc::CodeCommand
+  RUNDOC_ERB_BINDINGS = Hash.new { |h, k| h[k] = EmptyBinding.create }
+  RUNDOC_DEFAULT_ERB_BINDING = "default"
+end
