+++ lib/rundoc/code_command/print/erb.rb
@@ -1,21 +1,8 @@
 # frozen_string_literal: true
 
-require "erb"
-
-class EmptyBinding
-  def self.create
-    new.empty_binding
-  end
-
-  def empty_binding
-    binding
-  end
-end
+require_relative "../empty_binding"
 
 module Rundoc::CodeCommand
-  RUNDOC_ERB_BINDINGS = Hash.new { |h, k| h[k] = EmptyBinding.create }
-  RUNDOC_DEFAULT_ERB_BINDING = "default"
-
   class PrintERBArgs
     attr_reader :line, :binding_name
